# src/Encore.Api/wwwroot/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Encore",
    "version": "0.1.0",
    "description": "Event ticketing, as a modular monolith. Five modules, and since the Payments extraction two hosts: Encore.Api serves everything here except the three /internal/payments routes, which only Encore.Payments.Api maps. Paths that carry their own `servers` entry are the ones the host serving this page does not answer.\n\n**Actions, not resources.** The seat routes are `/hold`, `/release` and `/purchase` rather than a `/holds` collection: a hold is not an entity, it is two fields on the seat row, so it gets no URI of its own. All three are idempotent, which is what makes retrying a POST safe.\n\n**Branch on `reason`, not on status.** Every refusal about the state of the world is `409` with a machine-readable `reason` and a `retriable` flag. Only an addressed thing that is not there is `404`. A request that is wrong on its own terms is `400`. The statuses carry the class of failure; `reason` carries which one.\n\n**`X-Client-Id` is a claimed identity, not authentication.** Anyone can change it and reset their own hold cap. It stands in for the Identity module, which does not exist yet, which is also why no route returns 401 or 403.\n\nThis document is hand-maintained and was verified against a running instance. `Encore.Api` holds zero NuGet packages by design, so nothing generates this from endpoint metadata; a test fails when it and the mapped routes disagree."
  },
  "servers": [
...
  "components": {
    "securitySchemes": {
      "ClientId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Client-Id",
        "description": "A GUID identifying the calling client. NOT authentication — it is unverified and anyone may change it. Paste any GUID here to try the seat, order and payment routes."
      },
      "ServiceToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Service-Token",
        "description": "The shared secret between Orders and Payments. This one IS authentication, and it is the only thing guarding a call that can move money. Not a customer credential — the /internal routes take no X-Client-Id at all."
      }
    },
    "schemas": {
...
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem details. Module refusals add `reason` and `retriable`; framework refusals (404 for an unmatched route, 405, 415, a body that will not bind) carry only the standard members.",
        "properties": {
          "type": { "type": "string", "format": "uri" },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" },
          "instance": { "type": "string" },
          "reason": { "type": "string", "description": "Machine-readable refusal code. Branch on this." },
          "retriable": { "type": "boolean", "description": "Whether the identical request could plausibly succeed later." },
          "limit": { "type": "integer", "description": "Present on hold_cap_reached and too_many_seats." },
          "traceId": { "type": "string" }
        },
        "required": ["type", "title", "status"]
      },
...
      "PaymentResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "orderId": { "type": "string", "format": "uuid" },
          "status": { "type": "string", "enum": ["pending", "authorized", "captured", "declined", "timed_out", "voided"] },
          "amount": { "type": "number" },
          "currency": { "type": "string" },
          "gatewayReference": { "type": ["string", "null"] },
          "attemptedAt": { "type": "string", "format": "date-time" },
          "resolvedAt": { "type": ["string", "null"], "format": "date-time" }
        }
      },
      "AuthorizeAttemptRequest": {
        "type": "object",
        "description": "What Orders sends to open or reuse an authorisation. The client id is in the body rather than a header: the caller is Orders reporting a fact about the order, not a customer asserting who it is.",
        "required": ["orderId", "clientId", "amount", "currency"],
        "properties": {
          "orderId": { "type": "string", "format": "uuid" },
          "clientId": { "type": "string", "format": "uuid" },
          "amount": { "type": "number" },
          "currency": { "type": "string", "description": "ISO 4217." }
        }
      },
      "OrderAttemptRequest": {
        "type": "object",
        "description": "Keyed by order rather than by payment, because IOrderPayments is: the module owns which attempt an order currently means, and the one-live-attempt index picks it out.",
        "required": ["orderId", "clientId"],
        "properties": {
          "orderId": { "type": "string", "format": "uuid" },
          "clientId": { "type": "string", "format": "uuid" }
        }
      },
      "PaymentOutcomeResponse": {
        "type": "object",
        "description": "A successful service call. Branch on `outcome`, never on the status code — several statuses share a code, and the string is the half that is one-to-one with the contract vocabulary.",
        "required": ["outcome", "paymentId"],
        "properties": {
          "outcome": { "type": "string", "enum": ["authorized", "already_captured", "captured", "voided"] },
          "paymentId": { "type": "string", "format": "uuid" }
        }
      }
    },
    "responses": {
      "BadClientId": {
        "description": "The X-Client-Id header is missing or is not a GUID. No `reason`: the request is malformed rather than refused.",
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
      },
      "BadServiceToken": {
        "description": "The X-Service-Token header is missing, malformed or wrong. One refusal for all three, with `reason` service_token_invalid: telling an unauthenticated caller which of those it got tells it how to get closer.",
        "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } }
      }
    },
...
  },
  "paths": {
...
    "/payments/{paymentId}": {
      "get": {
        "tags": ["payments"],
        "summary": "Reads back what happened to a payment.",
        "security": [{ "ClientId": [] }],
        "parameters": [{ "name": "paymentId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "The attempt.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PaymentResponse" } } } },
          "400": { "$ref": "#/components/responses/BadClientId" },
          "404": { "description": "reason: payment_not_found.", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } }
        }
      }
    },
...
    "/internal/payments/authorize": {
      "servers": [
        { "url": "http://localhost:8081", "description": "payments-api. The monolith that serves this page does NOT map this route: MapPaymentsServiceApi is called only by Encore.Payments.Api." }
      ],
      "post": {
        "tags": ["payments-service"],
        "summary": "Opens or reuses this order's authorisation.",
        "description": "Idempotent per order: a repeat for an order that already has a live attempt returns that attempt rather than opening a second one, which is what makes a retry after a dropped response safe. Requires X-Service-Token; carries no X-Client-Id.",
        "security": [{ "ServiceToken": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthorizeAttemptRequest" } } } },
        "responses": {
          "200": { "description": "authorized, or already_captured.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PaymentOutcomeResponse" } } } },
          "401": { "$ref": "#/components/responses/BadServiceToken" },
          "402": { "description": "declined.", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
          "409": { "description": "concurrent_attempt_in_flight.", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
          "504": { "description": "timed_out. The attempt is recorded and the reconciler will settle it.", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } }
        }
      }
    },
    "/internal/payments/capture": {
      "servers": [
        { "url": "http://localhost:8081", "description": "payments-api. The monolith that serves this page does NOT map this route: MapPaymentsServiceApi is called only by Encore.Payments.Api." }
      ],
      "post": {
        "tags": ["payments-service"],
        "summary": "Takes the funds this order's attempt is holding.",
        "description": "Idempotent: capturing an already-captured attempt answers captured rather than failing.",
        "security": [{ "ServiceToken": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrderAttemptRequest" } } } },
        "responses": {
          "200": { "description": "captured.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PaymentOutcomeResponse" } } } },
          "401": { "$ref": "#/components/responses/BadServiceToken" },
          "409": { "description": "no_authorization.", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
          "504": { "description": "timed_out. The order is left awaiting capture and the next confirm resolves it.", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } }
        }
      }
    },
    "/internal/payments/void": {
      "servers": [
        { "url": "http://localhost:8081", "description": "payments-api. The monolith that serves this page does NOT map this route: MapPaymentsServiceApi is called only by Encore.Payments.Api." }
      ],
      "post": {
        "tags": ["payments-service"],
        "summary": "Releases what this order's attempt is holding.",
        "description": "Idempotent: voiding an already-voided attempt answers no_authorization rather than failing. A captured attempt refuses, because the money has already moved.",
        "security": [{ "ServiceToken": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrderAttemptRequest" } } } },
        "responses": {
          "200": { "description": "voided.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PaymentOutcomeResponse" } } } },
          "401": { "$ref": "#/components/responses/BadServiceToken" },
          "409": { "description": "already_captured, or no_authorization.", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
          "504": { "description": "timed_out.", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } }
        }
      }
    }
  }
}